Added: Git config
Add my standard Git config, as found in use.

Except, directives with anything personal to me or an organization I work
for have been omitted — user.email, url.insteadOf, etc.

Those can be placed in a separate file id_config, which isn't in this
public repository.

To handle that, also configure rcm to use 2 separate directory trees: the
default ~/.dotfiles/ (which is where these files are, and makes
bootstrapping easy) and also ~/.id/ (which is where .config/git/id_config
can be provided), so add a custom .rcrc as well.

Also, core.excludesFile has been removed, because that had a home
directory hard-coded into it. Instead, leave it at the default, which is
now ~/.config/git/ignore (and presumably wasn't when I started using
excludesFile), and rename the excludes file accordingly.