New: SSH config
Standard config I've been using (slightly pruned), with anything
server-specific removed.

Directives required for specific organizations' servers now in the ID
repository, and loaded with that Include local_config line — which I've
tried, and seems to be harmless when there isn't such a file.

The exception to specific host is GitHub, because that's public and used
in multiple places. Though maybe I should add a username to that?