New: Basic Vim config
My Vim RC files, as found live on my current laptop. No attempt to
sanitize them or preserve history beyond what's in the comments.

Other things in the various ~/.vim/ subdirectories not included yet,
because there's a lot of them and I'm pretty sure that I don't want all of
them.

Or for some of the things I do want, such as extensions, I want to install
them from their upstream location, not commit a copy of them into this
repository. Or maybe I do want them here, but as a Git submodule?