New: Basic Mate Terminal Config
Mate Terminal is a bit complicated. I'll want to set up more eventually,
but for now these are the basics that get it working well:

• Alt+F *doesn't* attempt to open the ‘File’ menu, meaning it works as
  expected in Bash to move forward by a word.

• F1 *doesn't* open the terminal's help window, meaning that it works as
  expected in Vim.

• Ctrl+Tab does navigate between tabs.

  I don't even know how to set this outside of dconf: it's something I
  wanted for years, and just chanced upon while viewing settings in
  dconf-editor. Setting the custom next/previous tab shortcuts doesn't
  allow pressing Tab; this is a specific boolean option just for this
  particular shortcut and nothing else.

  It also isn't clear why this isn't enabled by default anyway: it isn't
  like Alt+Tab does anything else.

But the major awkwardness here is that Mate Terminal doesn't use RC files,
instead having its config in dconf. So this config file has been generated
by doing:

  $ dconf dump /org/mate/terminal/

and then pruning it accordingly. And it needs a hook for loading it in
elsewhere.

Once the hook has done its thing, the file doesn't actually need to be
installed anywhere — so don't bother installing it. Instead have a
_manual/ subdirectory for config files that have to be loaded manually†.

dconf commands from: https://askubuntu.com/a/1302173/204161

† Well, not really manually, since I have a hook automating it. But I
  can't think of a better name.