Enhancement: Actual line-drawing characters
Switch output format to one which uses proper line-drawing characters
rather than repurposing Ascii symbols.

psql_unicode appears to be to psql as grid_fancy is to grid. And didn't
exist when I was last looking at these. It isn't mentioned in the
changelog, but it's in 3.5, and I want a recent version anyway for the
SSH tunnelling and sql-insert output format.