Fix: Git commit wrapping
When typing a Git commit message on the blank line at the top, don't start
wrapping it into the comment lines below it (starting with #). This
requires formatoptions having q in it.

I'm not actually sure why: the help for formatoptions suggests that having
t but not c should wrap text but not comments and that q makes comments
wrappable with the gq command, so it should already be doing the right
thing and also q shouldn't fix it. But it seems that it does, and I
already had this in this file, so it must be something that I discovered
previously.

I'm guessing this isn't broken as supplied but has been by a default
formatoptions I set in vimrc which now needs overriding for this filetype.
But I haven't checked.

Also brings in some other suitable settings for Git commits, which in turn
requires those for the human filetype.