Refactoring: share isn't dotted
I've re-read the docs for file patterns and understand them now. The bit
before the colon is just for distinguishing ~/.dotfiles/ from ~/.id/ or
similar, which isn't something I think I need to do. The rest is a
relative path from within the dotfiles directory.

So share is just share, in order to be dotless.