New: Vim Easy Align plug-in
Bring in the Vim Easy Align plug-in (for easily aligning things).

Don't bother copying the entire codebase; it's a Git repository, so
include it as a Git submodule.

Place it in the appropriate ~/.vim/ subdirectory to load it as a package.
Then when installing it with RCM, link the ~/.vim/pack/EasyAlign
subdirectory, because nothing inside that (or indeed any other package)
should change, and it's cleaner to link in the entire plug-in than each of
its files separately. (Especially if the plug-in ever deletes, renames, or
moves files, I suppose.)