Fix: Vim EasyAlign package actually loads
For Vim to find a package and load it when it starts, it needs to be in a
directory nested 2 levels deeper.

Fortunately this was any easy move: I just did:

  $ mkir vim/pack/EasyAlign/start
  $ git mv vim/pack/vim-easy-align !$

and Git was aware enough to make the corresponding change to .gitmodules
as well.

rcrc already takes care of this, as the symlink can still be at the
~/.vim/pack/EasyAlign level.