Enhancement: Generate Vim package help tags
After installing Vim packages, generate help tags for their documentation.

This saves the tags in tags files in the doc/ subdirectories. Both
packages I'm already using have doc/tags in their own .gitignore files, so
I don't need to do anything extra for that.

Use ex rather than vim to run the command because that doesn't mess up the
terminal: first complaining that it isn't a TTY, then leaving it in an
unusable state that requires using reset to make it workable again.