Enhancement: Load GitGutter when appropriate
Trigger loading the GitGutter package when the file being created or
edited is in a directory being tracked by Git — but still default to its
being off, so it doesn't cause clutter when editing files elsewhere.

This replaces a previous mechanism, a fake plug-in called gitgutter_load,
which my gvimrc was attempting to invoke, but which actually just led to
an error because that plug-in wasn't included in this repository. Now that
Vim has built-in packages support, including ./opt/ subdirectories and the
:packadd command, this way is superior.