Enhancement: gy to yank filename into X selection
Mapping to yank the current filename to the X selection with gy. Chosen
because Ctrl+G shows the filename and y does yanking.

Do it as its own plugin, since this isn't specific to me so could
(theoretically) be shared and installed by others.

Put the plugin in a package, because that seems like a reasonable way of
making it shareable (possibly?). Create a new subdirectory in pack/ for
packages that are mine rather than from elsewhere and downloaded as Git
submodules. Though, really, it still feels like there's at least one layer
too many here and it's all a bit arbitrary.

Using a package also means that the existing rcup hook will generate the
tags file, to make :h gy work. Ignore all doc/tags files in this and any
future packages I add in the mine directory.

My first vim9script! Largely based on :h write-plugin — though with one
level less indirection in the mapping, since this doesn't have a menu item
to share with the keystrokes.