Fix: Silently do nothing without Vim9 Script
This plug-in is written in Vim9 Script. So don't try to run it on older
versions of Vim that don't have this language.

It's an optional feature, so missing it out isn't terrible (and it's
unusual for me to encounter a pre-Vim9 Script Vim these days), but it's
better to silently do nothing that to display a bunch of error messages
from trying to interpret Vim9 Script commands (and comments) as legacy Vim
Script.