New: Ubuntu theme at log-in
When logging in on a desktop environment, play the Ubuntu theme tune — the
original Ubuntu start-up sound.

Download the theme and ensure the paplay command is installed.

Playing the start-up sound was disabled by Ubuntu 12.04. At some point
since then the installed sound (still not played by default) was changed.

Based on:
https://www.omgubuntu.co.uk/2024/03/enable-old-ubuntu-login-sound

Except don't bother trying to patch the theme or make this an ‘official’
start-up sound; just have it as a stand-alone command which plays it,
since that's simpler.