Enhancement: Avoid bright yellow and cyan
Mate Terminal defaults to using brighter colours for bold. Bright yellow
and bright cyan are hard to read on a while or pale background colour, so
avoid using them.

Instead, use the normal shades of those colours but reversed (pale text on
a block of colour), so they are still distinguished from the non-main
branches.