Fix: Playing Ubuntu theme when logging in
Actually download the right file, and actually play the file that was
downloaded. Commit 4f2d6842 messed both of those up, which then balanced
each other out, because it was playing a file I'd downloaded elsewhere.

The original URL was the HTML page for playing or downloading the music
file, not the file itself.