Enhancement: No per-tab close buttons
Get rid of close buttons on individual tabs, since they get in the way.
Even just having one on the current tab reduces its useful area and leads
to its sometimes getting clicked on by mistake.

Middle-click already works anywhere to close a tab, and I have a close
button at the right of the tab bar.

This involves creating userChrome.css, linking it into the random profile
directory, and the preference for telling Firefox to look for
userChrome.css. Fortunately the preference can easily be added to user.js
created for something else, and the chrome directory can be linked in like
user.js is with a little modification to the script.