Enhancement: Minimum font size
Restrict how small the font can go on pages. Set based on UK Game Shows,
where this was hard to read by default:
http://www.ukgameshows.com/ukgs/Popmaster

However, this may be too big for what's supposed to actually be ‘small
print’. It'd be better to have a setting that only applies to the main
body text on a page, but let's try this for now.