Fix: Ensure Vim installed before .vimrc
Annoyingly, the minimal Vim providing the vi command in default Ubuntu
looks at .vimrc files — which I feel like it shouldn't, for something
invoked as vi.

And, further annoyingly, on given an invalid command in .vimrc, it fails
with an error saying that it isn't supported in that version.

So deploying my custom .vimrc breaks the default vi command. Stop that
happening by installing a proper Vim before then.