Fix: Work without $PAGER being set
Ack's output was entirely blank when this .ackrc was installed but $PAGER
was empty.

Allow for that by defaulting to cat. (Don't default to less, because it
can't cope with the colour codes until $LESS has been set with appropriate
options.)

.bashrc in this repository sets $LESS. But at the point when rcup has just
installed all these files, ack instantly starts using .ackrc, but Bash
only uses .bashrc on starting a shell.

(A post-up hook that sources .bashrc wouldn't help, because that would
itself run in a subshell.)