New: Additional applications to install
Applications and other packages I want installing by default.

Some of the echo and $pkg bits could probably be usefully abstracted out
into a function, but they aren't quite the same; it was useful to do it
like this first in order to be able to see the common parts.