Enhancement: Super+N for VPN toggling
Have Super+N toggle the VPN on/off, by running vpn_toggle. This requires
the recent improvement where the VPN name doesn't need to be specified, if
there's only one defined in NetworkManager.

Gnome had Super+N defined to focus the active notification. I haven't
managed to work out what this does (what makes a notification active, and
what focusing it would involve), but I've moved put it on Shift+Super+N,
so it's still available if it turns out to be useful.