Enhancement: Super+O to see track playing
It turns out that ‘show the notification list’ is actually useful, because
at the top it shows the track currently playing. I'm used to that being on
Super+O, which seems to be available, so move it to that.

It's a shame not to leave this on its standard keystroke, but I'm too used
to Super+V being Vim to change that. And there's nothing about ‘V’ which
indicates ‘notifications’ or ‘now playing’ or ‘message tray’ anyway.