Config: Gnome Terminal
Some things were unreadable in the default Gnome Terminal with a dark
background, so switch to the ‘Gnome Light’ colour scheme.

While here, also switch the interface to follow the system theme, so that
matches other applications. It was initially set to ‘always be dark’,
which is curious — somebody who wants it dark can always choose a dark
theme.

And make Ctrl+Tab and Ctrl+Shift+Tab work for switching tabs. Also
curiously, these can't be set through the preferences interface (and now
having set them like this, that shows ‘Invalid’ for these), but specifying
Tab like this works fine.

Also have Alt+9 be the rightmost tab (like in Firefox), so that it always
works and I don't have to count them. (Though unlike in Firefox, this is
because I generally have fewer than 9 tabs, not more.)