Config: Zathura for PDF viewing
Specify Zathura as the default PDF viewer, in Gnome.

It seems some applications write to ~/.config/mimeapps.list (such as
installing the Firefox beta, or making it my default browser?). Avoid the
clash by writing to the Gnome-specific mime list file, which overrides the
general one.

It isn't that I want this setting to be specific to Gnome; it's just a
convenient place to put it.