Fix: Ignore ~/ when checking if command exists
When checking whether a command exists, to determine whether its package
is installed, ignore commands in ~/bin/ and similar. Specifically, before
doing the check, remove any paths starting $HOME from the path.

The pattern has to be ^$HOME not ^~, because the tilde doesn't get
expanded in the latter.

This is needed for a forthcoming change to add an, which already has a
wrapper in ~/bin/an which doesn't work because the system an isn't
installed. Having the check find ~/bin/an would be pointless.